+++ CHANGELOG.md
@@ -1,9 +1,22 @@
 <!-- markdownlint-disable MD024 -->
 # Changelog
 
-## [1.0.0] - 2026-02-01
+
+## [Unreleased]
+
+### Added
+
+- #2 Add DSC resource for project permissions## [1.0.0] - 2026-02-01
 
 ### Added
 
 - Initial release. Add DSC class for:  Users, Projects, Groups
+
+## [1.0.1] - 2026-03-13
+
+### Added
+
+- Add class: ProjectUserPermissionResource
+- Add class: ProjectGroupPermissionResource
+